% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_timepoint}
\alias{parse_timepoint}
\title{parse a timepoint}
\usage{
parse_timepoint(timepoint)
}
\arguments{
\item{timepoint}{timepoint object as in zc_readings object}
}
\description{
parse a timepoint
}
