% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zentra.R
\name{zc_token}
\alias{zc_token}
\title{retrieve authentication token from zentracloud}
\usage{
zc_token(
  username = Sys.getenv("ZC_USERNAME"),
  password = Sys.getenv("ZC_PASSWORD")
)
}
\arguments{
\item{username}{username. Read by default from environment variable \code{ZC_USERNAME}}

\item{password}{password. Read by default from environment variable \code{ZC_PASSWORD}}
}
\value{
api token
}
\description{
retrieve authentication token from zentracloud
}
\examples{
\dontrun{
token <- zc_token("xxxx", "yyyyy")
}
}
